refactor(RPG Maker Structs): :art: Begin to refactor RMXP types to make it mroe ergonomic to use
