+++ rmxp-types/Cargo.toml
@@ -12,8 +12,10 @@ bytemuck = { version = "1.13.1", features = ["extern_crate_alloc"] }
 num-traits = "0.2"
 num-derive = "0.3"
 slab = { version = "0.4", features = ["serde"] }
-strum = { version = "0.24.1", features = ["derive"] }
+strum = { version = "0.25", features = ["derive"] }
 paste = "1.0"
+camino = "1.1"
+num_enum = "0.6"
 
 rand = "0.8.5"
 flate2 = "1.0"
