+++ rmxp-types/src/lib.rs
@@ -1,12 +1,35 @@
 #![feature(min_specialization)]
 
-mod nil_padded;
-mod parameter_type;
+// Editor specific types
+pub mod rmxp;
+
+// Shared structs with the same layout
+mod shared;
+
 mod rgss_structs;
 
-pub mod rmxp_structs;
+mod helpers;
 
-pub use nil_padded::NilPadded;
-pub use parameter_type::ParameterType;
+pub use helpers::*;
 pub use rgss_structs::{Color, Table1, Table2, Table3, Tone};
-pub use rmxp_structs as rpg;
+
+pub mod rpg {
+    pub use crate::rmxp::*;
+    pub use crate::shared::*;
+
+    pub type Actors = Vec<Actor>;
+    pub type Animations = Vec<Animation>;
+    pub type Armors = Vec<Armor>;
+    pub type Classes = Vec<Class>;
+    pub type CommonEvents = Vec<CommonEvent>;
+    pub type Enemies = Vec<Enemy>;
+    pub type Items = Vec<Item>;
+    pub type MapInfos = std::collections::HashMap<i32, MapInfo>;
+    pub type Skills = Vec<Skill>;
+    pub type States = Vec<State>;
+    pub type Tilesets = Vec<Tileset>;
+    pub type Troops = Vec<Troop>;
+    pub type Weapons = Vec<Weapon>;
+}
+
+pub type Path = Option<camino::Utf8PathBuf>;
