+++ rmxp-types/src/rmxp/mod.rs
@@ -1,0 +1,44 @@
+// Copyright (C) 2023 Lily Lyons
+//
+// This file is part of Luminol.
+//
+// Luminol is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Luminol is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Luminol.  If not, see <http://www.gnu.org/licenses/>.
+
+mod actor;
+mod animation;
+mod armor;
+mod class;
+mod enemy;
+mod item;
+mod map;
+mod skill;
+mod state;
+mod system;
+mod tileset;
+mod troop;
+mod weapon;
+
+pub use actor::*;
+pub use animation::*;
+pub use armor::*;
+pub use class::*;
+pub use enemy::*;
+pub use item::*;
+pub use map::*;
+pub use skill::*;
+pub use state::*;
+pub use system::*;
+pub use tileset::*;
+pub use troop::*;
+pub use weapon::*;
