+++ rmxp-types/src/shared/event.rs
@@ -1,0 +1,158 @@
+// Copyright (C) 2023 Lily Lyons
+//
+// This file is part of Luminol.
+//
+// Luminol is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Luminol is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Luminol.  If not, see <http://www.gnu.org/licenses/>.
+use crate::{optional_path, rpg::MoveRoute, ParameterType, Path};
+
+#[derive(Debug, serde::Deserialize, serde::Serialize, Clone)]
+#[serde(rename = "RPG::Event")]
+pub struct Event {
+    pub id: usize,
+    pub name: String,
+    pub x: i32,
+    pub y: i32,
+    pub pages: Vec<EventPage>,
+}
+
+impl Event {
+    #[must_use]
+    pub fn new(x: i32, y: i32, id: usize) -> Self {
+        Self {
+            id,
+            name: format!("EV{id:0>3}"),
+            x,
+            y,
+            pages: vec![EventPage::default()],
+        }
+    }
+}
+
+#[derive(Debug, serde::Deserialize, serde::Serialize, Clone)]
+#[serde(rename = "RPG::Event::Page")]
+pub struct EventPage {
+    pub condition: EventCondition,
+    pub graphic: Graphic,
+    pub move_type: usize,
+    pub move_speed: usize,
+    pub move_frequency: usize,
+    pub move_route: MoveRoute,
+    pub walk_anime: bool,
+    pub step_anime: bool,
+    pub direction_fix: bool,
+    pub through: bool,
+    pub always_on_top: bool,
+    pub trigger: i32,
+    pub list: Vec<EventCommand>,
+}
+
+impl Default for EventPage {
+    fn default() -> Self {
+        Self {
+            condition: EventCondition::default(),
+            graphic: Graphic::default(),
+            move_type: 0,
+            move_speed: 3,
+            move_frequency: 3,
+            move_route: MoveRoute::default(),
+            walk_anime: true,
+            step_anime: false,
+            direction_fix: false,
+            through: false,
+            always_on_top: false,
+            trigger: 0,
+            list: vec![],
+        }
+    }
+}
+
+#[derive(Debug, serde::Deserialize, serde::Serialize, Clone)]
+#[serde(rename = "RPG::Event::Page::Graphic")]
+pub struct Graphic {
+    pub tile_id: i32,
+    #[serde(with = "optional_path")]
+    pub character_name: Path,
+    pub character_hue: i32,
+    pub direction: i32,
+    pub pattern: i32,
+    pub opacity: i32,
+    pub blend_type: i32,
+}
+
+impl Default for Graphic {
+    fn default() -> Self {
+        Self {
+            tile_id: 0,
+            character_name: None,
+            character_hue: 0,
+            direction: 2,
+            pattern: 0,
+            opacity: 255,
+            blend_type: 0,
+        }
+    }
+}
+
+#[derive(Debug, serde::Deserialize, serde::Serialize, Clone)]
+#[serde(rename = "RPG::Event::Page::Condition")]
+pub struct EventCondition {
+    pub switch1_valid: bool,
+    pub switch2_valid: bool,
+    pub variable_valid: bool,
+    pub self_switch_valid: bool,
+    pub switch1_id: usize,
+    pub switch2_id: usize,
+    pub variable_id: usize,
+    pub variable_value: i32,
+    pub self_switch_ch: String,
+}
+
+impl Default for EventCondition {
+    fn default() -> Self {
+        Self {
+            switch1_valid: false,
+            switch2_valid: false,
+            variable_valid: false,
+            self_switch_valid: false,
+            switch1_id: 1,
+            switch2_id: 1,
+            variable_id: 1,
+            variable_value: 0,
+            self_switch_ch: "A".to_string(),
+        }
+    }
+}
+
+#[derive(Default, Debug, serde::Deserialize, serde::Serialize, Clone)]
+#[serde(rename = "RPG::CommonEvent")]
+pub struct CommonEvent {
+    pub id: usize,
+    pub name: String,
+    pub trigger: usize,
+    pub switch_id: usize,
+    pub list: Vec<EventCommand>,
+}
+
+#[derive(Debug, serde::Deserialize, serde::Serialize, Clone)]
+#[allow(missing_docs)]
+#[serde(rename = "RPG::EventCommand")]
+pub struct EventCommand {
+    pub code: u16,
+    pub indent: usize,
+    pub parameters: Vec<ParameterType>,
+
+    #[serde(default = "rand::random")]
+    #[serde(skip)]
+    pub guid: u16,
+}
