+++ rmxp-types/src/shared/mod.rs
@@ -1,0 +1,28 @@
+// Copyright (C) 2023 Lily Lyons
+//
+// This file is part of Luminol.
+//
+// Luminol is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Luminol is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Luminol.  If not, see <http://www.gnu.org/licenses/>.
+
+mod audio_file;
+mod event;
+mod mapinfo;
+mod move_route;
+mod script;
+
+pub use audio_file::*;
+pub use event::*;
+pub use mapinfo::*;
+pub use move_route::*;
+pub use script::*;
