+++ src/cache/data.rs
@@ -308,18 +308,18 @@ impl Cache {
         let scripts = AtomicRefCell::new(scripts);
 
         *self.state.borrow_mut() = State::Loaded {
-            actors: AtomicRefCell::new(NilPadded::from(vec![rpg::Actor::default()])),
-            animations: AtomicRefCell::new(NilPadded::from(vec![rpg::Animation::default()])),
-            armors: AtomicRefCell::new(NilPadded::from(vec![rpg::Armor::default()])),
-            classes: AtomicRefCell::new(NilPadded::from(vec![rpg::Class::default()])),
-            common_events: AtomicRefCell::new(NilPadded::from(vec![rpg::CommonEvent::default()])),
-            enemies: AtomicRefCell::new(NilPadded::from(vec![rpg::Enemy::default()])),
-            items: AtomicRefCell::new(NilPadded::from(vec![rpg::Item::default()])),
-            skills: AtomicRefCell::new(NilPadded::from(vec![rpg::Skill::default()])),
-            states: AtomicRefCell::new(NilPadded::from(vec![rpg::State::default()])),
-            tilesets: AtomicRefCell::new(NilPadded::from(vec![rpg::Tileset::default()])),
-            troops: AtomicRefCell::new(NilPadded::from(vec![rpg::Troop::default()])),
-            weapons: AtomicRefCell::new(NilPadded::from(vec![rpg::Weapon::default()])),
+            actors: AtomicRefCell::new(vec![rpg::Actor::default()]),
+            animations: AtomicRefCell::new(vec![rpg::Animation::default()]),
+            armors: AtomicRefCell::new(vec![rpg::Armor::default()]),
+            classes: AtomicRefCell::new(vec![rpg::Class::default()]),
+            common_events: AtomicRefCell::new(vec![rpg::CommonEvent::default()]),
+            enemies: AtomicRefCell::new(vec![rpg::Enemy::default()]),
+            items: AtomicRefCell::new(vec![rpg::Item::default()]),
+            skills: AtomicRefCell::new(vec![rpg::Skill::default()]),
+            states: AtomicRefCell::new(vec![rpg::State::default()]),
+            tilesets: AtomicRefCell::new(vec![rpg::Tileset::default()]),
+            troops: AtomicRefCell::new(vec![rpg::Troop::default()]),
+            weapons: AtomicRefCell::new(vec![rpg::Weapon::default()]),
 
             mapinfos,
             maps,
